% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_UI_functions.R
\name{add_plot_maximize_observer}
\alias{add_plot_maximize_observer}
\title{Add a box maximization observer to automatically resize a plot in that box.}
\usage{
add_plot_maximize_observer(input, box_id, plot_name, non_max_height = "400px")
}
\arguments{
\item{input}{The input of a shiny app session.}

\item{box_id}{The shiny ID of the box to observe.}

\item{plot_name}{The shiny ID of the plot to resize.}

\item{non_max_height}{The height that the graph should be when the box is
not maximized. Defaults to "400px".
@import plotly}
}
\description{
Add a box maximization observer to automatically resize a plot in that box.
}
