% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar_chart.R
\name{mod_bar_chart_server}
\alias{mod_bar_chart_server}
\title{bar_chart Server Functions}
\usage{
mod_bar_chart_server(
  id,
  bar_chart_values = tasks_not_run_by_study_count()$n,
  bar_chart_labels = tasks_not_run_by_study_count()$Study.Name,
  x_axis_name = "Study Name",
  y_axis_name = "Total # Tasks"
)
}
\description{
bar_chart Server Functions
}
