% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_bar_chart.R
\name{mod_bar_chart_ui}
\alias{mod_bar_chart_ui}
\title{bar_chart UI Function}
\usage{
mod_bar_chart_ui(id, bar_chart_title = "Pending Deliverables Across Studies")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
