% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kpi_card_overview.R
\name{mod_kpi_card_overview_ui}
\alias{mod_kpi_card_overview_ui}
\title{kpi_card_overview UI Function}
\usage{
mod_kpi_card_overview_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\examples{
ui <- bs4Dash::dashboardPage(
  bs4Dash::dashboardHeader(),
  bs4Dash::dashboardSidebar(disable = TRUE),
  bs4Dash::dashboardBody(
    fluidRow(
      mod_kpi_card_overview_ui("kpi_card_example")
    )
  )
)
server <- function(input, output) {
  mod_kpi_card_overview_server("kpi_card_example")
}

shiny::shinyApp(ui, server)

}
