% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kpi_card.R
\name{mod_kpi_card_ui}
\alias{mod_kpi_card_ui}
\title{kpi_card UI Function}
\usage{
mod_kpi_card_ui(id)
}
\description{
This module will create value box for key indicators of task analysis like total number of task with no due' dates etc.
}
\details{
@param id,input,output,session Internal parameters for {shiny}.

 @importFrom shiny NS tagList
}
\examples{

ui <- bs4Dash::dashboardPage(
  bs4Dash::dashboardHeader(),
  bs4Dash::dashboardSidebar(disable = TRUE),
  bs4Dash::dashboardBody(
    fluidRow(
      mod_kpi_card_ui("kpi_card_example")
    )
  )
)

server <- function(input, output) {
  mod_kpi_card_server("kpi_card_example")
}
shiny::shinyApp(ui, server)

}
