% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pie_chart.R
\name{mod_pie_chart_ui}
\alias{mod_pie_chart_ui}
\title{pie_chart UI Function}
\usage{
mod_pie_chart_ui(id, pie_chart_title)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module.
}
\examples{
ui <- bs4Dash::dashboardPage(
  bs4Dash::dashboardHeader(),
  bs4Dash::dashboardSidebar(disable = TRUE),
  bs4Dash::dashboardBody(
    fluidRow(
      mod_pie_chart_ui("pie_chart_example", pie_chart_title = "Pie Chart")
    )
  )
)
server <- function(input, output) {
  mod_pie_chart_server("pie_chart_example", pie_chart_values = task_completion_rate()$n, pie_chart_labels = task_completion_rate()$Completed)
}

shiny::shinyApp(ui, server)

}
