% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pie_charts_overview.R
\name{mod_pie_charts_overview_ui}
\alias{mod_pie_charts_overview_ui}
\title{pie_charts_overview UI Function}
\usage{
mod_pie_charts_overview_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module. right now not in use
}
\examples{
ui <- bs4Dash::dashboardPage(
  bs4Dash::dashboardHeader(),
  bs4Dash::dashboardSidebar(disable = TRUE),
  bs4Dash::dashboardBody(
    fluidRow(
      mod_pie_charts_overview_ui("pie_charts_example")
    )
  )
)
server <- function(input, output) {
  mod_pie_charts_overview_server("pie_charts_example")
}

shiny::shinyApp(ui, server)
}
