% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{task_summary_data_fun}
\alias{task_summary_data_fun}
\title{task_summary_data
Read task summary data from data for reporting excel file for calculations}
\usage{
task_summary_data_fun()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
A number.
}
\description{
task_summary_data
Read task summary data from data for reporting excel file for calculations
}
